requirements.txt
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
segment-anything>=1.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tqdm>=4.62.0
accelerate>=0.4.0